backend/requirements.txt
apiclient==1.0.4
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.6.3
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
